MiniProject_2/requirements.txt
getpass4
openai
python-dotenv